Initial (buggy) port to standalone apps
