+++ nbgrader/apps/assignapp.py
@@ -1,36 +1,23 @@
-import os
-import re
-from textwrap import dedent
+import sys
 
-from traitlets import List, Bool, observe, default
+from traitlets import default
 
-from ..api import Gradebook, MissingEntry
-from .baseapp import (
-    BaseNbConvertApp, NbGraderException, nbconvert_aliases, nbconvert_flags)
-from ..preprocessors import (
-    IncludeHeaderFooter,
-    ClearSolutions,
-    LockCells,
-    ComputeChecksums,
-    SaveCells,
-    CheckCellMetadata,
-    ClearOutput,
-    ClearHiddenTests,
-)
+from .baseapp import NbGrader, nbgrader_aliases, nbgrader_flags
+from ..converters import BaseConverter, Assign, NbGraderException
 
 aliases = {}
-aliases.update(nbconvert_aliases)
+aliases.update(nbgrader_aliases)
 del aliases['student']
 aliases.update({
 })
 
 flags = {}
-flags.update(nbconvert_flags)
+flags.update(nbgrader_flags)
 flags.update({
     'no-db': (
         {
             'SaveCells': {'enabled': False},
-            'AssignApp': {'no_database': True}
+            'Assign': {'no_database': True}
         },
         "Do not save information into the database."
     ),
@@ -44,13 +31,17 @@ flags.update({
         "Do not validate or modify cell metatadata."
     ),
     'create': (
-        {'AssignApp': {'create_assignment': True}},
+        {'Assign': {'create_assignment': True}},
         "Create an entry for the assignment in the database, if one does not already exist."
+    ),
+    'force': (
+        {'BaseConverter': {'force': True}},
+        "Overwrite an assignment/submission if it already exists."
     ),
 })
 
 
-class AssignApp(BaseNbConvertApp):
+class AssignApp(NbGrader):
 
     name = u'nbgrader-assign'
     description = u'Produce the version of an assignment to be released to students.'
@@ -106,126 +97,40 @@ class AssignApp(BaseNbConvertApp):
 
         """
 
-    create_assignment = Bool(
-        False,
-        help=dedent(
-            """
-            Whether to create the assignment at runtime if it does not
-            already exist.
-            """
-        )
-    ).tag(config=True)
+    @default("classes")
+    def _classes_default(self):
+        classes = super(AssignApp, self)._classes_default()
+        classes.extend([BaseConverter, Assign])
+        return classes
 
-    no_database = Bool(
-        False,
-        help=dedent(
-            """
-            Do not save information about the assignment into the database.
-            """
-        )
-    ).tag(config=True)
+    def _load_config(self, cfg, **kwargs):
+        if 'AssignApp' in cfg:
+            self.log.warning(
+                "Use Assign in config, not AssignApp. Outdated config:\n%s",
+                '\n'.join(
+                    'AssignApp.{key} = {value!r}'.format(key=key, value=value)
+                    for key, value in cfg.AssignApp.items()
+                )
+            )
+            cfg.Assign.merge(cfg.AssignApp)
+            del cfg.AssignApp
 
-    @default("permissions")
-    def _permissions_default(self):
-        return 644
+        super(AssignApp, self)._load_config(cfg, **kwargs)
 
-    @property
-    def _input_directory(self):
-        return self.coursedir.source_directory
+    def start(self):
+        super(AssignApp, self).start()
 
-    @property
-    def _output_directory(self):
-        return self.coursedir.release_directory
+        if len(self.extra_args) > 1:
+            self.fail("Only one argument (the assignment id) may be specified")
+        elif len(self.extra_args) == 1 and self.coursedir.assignment_id != "":
+            self.fail("The assignment cannot both be specified in config and as an argument")
+        elif len(self.extra_args) == 0 and self.coursedir.assignment_id == "":
+            self.fail("An assignment id must be specified, either as an argument or with --assignment")
+        elif len(self.extra_args) == 1:
+            self.coursedir.assignment_id = self.extra_args[0]
 
-    export_format = 'notebook'
-
-    preprocessors = List([
-        IncludeHeaderFooter,
-        LockCells,
-        ClearSolutions,
-        ClearOutput,
-        CheckCellMetadata,
-        ComputeChecksums,
-        SaveCells,
-        ClearHiddenTests,
-        ComputeChecksums,
-        CheckCellMetadata,
-    ])
-    # NB: ClearHiddenTests must come after ComputeChecksums and SaveCells.
-    # ComputerChecksums must come again after ClearHiddenTests.
-
-    def build_extra_config(self):
-        extra_config = super(AssignApp, self).build_extra_config()
-        extra_config.CourseDirectory.student_id = '.'
-        extra_config.CourseDirectory.notebook_id = '*'
-        return extra_config
-
-    def _clean_old_notebooks(self, assignment_id, student_id):
-        with Gradebook(self.coursedir.db_url) as gb:
-            assignment = gb.find_assignment(assignment_id)
-            regexp = re.escape(os.path.sep).join([
-                self._format_source("(?P<assignment_id>.*)", "(?P<student_id>.*)", escape=True),
-                "(?P<notebook_id>.*).ipynb"
-            ])
-
-            # find a set of notebook ids for new notebooks
-            new_notebook_ids = set([])
-            for notebook in self.notebooks:
-                m = re.match(regexp, notebook)
-                if m is None:
-                    raise NbGraderException("Could not match '%s' with regexp '%s'", notebook, regexp)
-                gd = m.groupdict()
-                if gd['assignment_id'] == assignment_id and gd['student_id'] == student_id:
-                    new_notebook_ids.add(gd['notebook_id'])
-
-            # pull out the existing notebook ids
-            old_notebook_ids = set(x.name for x in assignment.notebooks)
-
-            # no added or removed notebooks, so nothing to do
-            if old_notebook_ids == new_notebook_ids:
-                return
-
-            # some notebooks have been removed, but there are submissions associated
-            # with the assignment, so we don't want to overwrite stuff
-            if len(assignment.submissions) > 0:
-                msg = "Cannot modify existing assignment '%s' because there are submissions associated with it" % assignment
-                self.error(msg)
-                raise NbGraderException(msg)
-
-            # remove the old notebooks
-            for notebook_id in (old_notebook_ids - new_notebook_ids):
-                self.log.warning("Removing notebook '%s' from the gradebook", notebook_id)
-                gb.remove_notebook(notebook_id, assignment_id)
-
-    def init_assignment(self, assignment_id, student_id):
-        super(AssignApp, self).init_assignment(assignment_id, student_id)
-
-        # try to get the assignment from the database, and throw an error if it
-        # doesn't exist
-        if not self.no_database:
-            assignment = {}
-            for a in self.db_assignments:
-                if a['name'] == assignment_id:
-                    assignment = a.copy()
-                    break
-
-            if assignment or self.create_assignment:
-                if 'name' in assignment:
-                    del assignment['name']
-                self.log.info("Updating/creating assignment '%s': %s", assignment_id, assignment)
-                with Gradebook(self.coursedir.db_url) as gb:
-                    gb.update_or_create_assignment(assignment_id, **assignment)
-
-            else:
-                with Gradebook(self.coursedir.db_url) as gb:
-                    try:
-                        gb.find_assignment(assignment_id)
-                    except MissingEntry:
-                        msg = "No assignment called '%s' exists in the database" % assignment_id
-                        self.error(msg)
-                        raise NbGraderException(msg)
-
-            # check if there are any extra notebooks in the db that are no longer
-            # part of the assignment, and if so, remove them
-            if self.coursedir.notebook_id == "*":
-                self._clean_old_notebooks(assignment_id, student_id)
+        converter = Assign(coursedir=self.coursedir, parent=self)
+        try:
+            converter.start()
+        except NbGraderException:
+            sys.exit(1)
