+++ nbgrader/converters/__init__.py
@@ -1,0 +1,12 @@
+from .base import BaseConverter, NbGraderException
+from .assign import Assign
+from .autograde import Autograde
+from .feedback import Feedback
+
+__all__ = [
+    "BaseConverter",
+    "NbGraderException",
+    "Assign",
+    "Autograde",
+    "Feedback"
+]
